floats eliminated

+++ DecisionUnitInterface/src/decisionunit/itf/sensors/clsEnergyConsumption.java
@@ -17,7 +17,7 @@ public class clsEnergyConsumption extends clsSensorHomeostasis {
 	public String logXML() {
 		String logEntry = "";
 		
-		logEntry += addXMLTag("Energy", new Float(mrEnergy).toString()); 
+		logEntry += addXMLTag("Energy", new Double(mrEnergy).toString()); 
 
 		return addXMLTag(logEntry);
 	}
